Move visitors metric widget out of toolkit
