+++ projects/packages/premium-analytics/packages/widgets-toolkit/src/index.ts
@@ -7,6 +7,7 @@ export {
 	ComparativeLineChart,
 	DonutChart,
 	Legend,
+	ReportMetricWidget,
 	WidgetRoot,
 	useWidgetRootContext,
 	type DonutChartData,
@@ -87,7 +88,6 @@ export {
 	OrdersFulfillmentWidget,
 	SalesByCouponWidget,
 	TotalReturnsWidget,
-	VisitorMetricWidget,
 	VisitorsByLocationWidget,
 	SalesByDeviceWidget,
 	BookingsByDeviceWidget,
