+++ projects/packages/premium-analytics/packages/widgets-toolkit/src/widgets/index.ts
@@ -1,6 +1,5 @@
 export { MetricComparisonWidget } from './metric-comparison';
 export { OrderMetricWidget, BookingOrderMetricWidget } from './order-metric';
-export { VisitorMetricWidget } from './visitor-metric';
 export { SalesByCouponWidget } from './sales-by-coupon';
 export { ConversionRateWidget, BookingConversionRateWidget } from './conversion-rate';
 export {
