--- projects/packages/premium-analytics/packages/widgets-toolkit/src/widgets/visitor-metric/index.ts
@@ -1,1 +1,0 @@
-export { VisitorMetricWidget } from './widget-visitor-metric';
